Added a new faded vertical scroll class for nutritional content details and daily values in food detail view.
